+++ src/app/components/directive-collection.ts
@@ -1,4 +1,4 @@
-import { TerraDataTableContextMenuDirective } from './tables/data-table/context-menu/directive/terra-data-table-context-menu.directive';
+import { TerraDataTableContextMenuDirective } from './tables/data-table/context-menu/terra-data-table-context-menu.directive';
 import { Type } from '@angular/core';
 import { TerraTwoColumnsContainerDirective } from './layouts/column-container/two-columns/terra-two-columns-container.directive';
 import { FixedHeaderDirective } from './tables/fixed-header/fixed-header.directive';
