+++ src/app/components/tables/data-table/context-menu/terra-data-table-context-menu.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { TerraDataTableContextMenuService } from './service/terra-data-table-context-menu.service';
+import { TerraDataTableContextMenuService } from './terra-data-table-context-menu.service';
 import { TerraDataTableContextMenuEntryInterface } from './data/terra-data-table-context-menu-entry.interface';
 import { TerraBaseData } from '../../../data/terra-base.data';
