+++ src/app/components/tables/data-table/context-menu/terra-data-table-context-menu.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Subject } from 'rxjs/Rx';
-import { TerraDataTableContextMenuEntryInterface } from '../data/terra-data-table-context-menu-entry.interface';
-import { TerraBaseData } from '../../../../data/terra-base.data';
+import { TerraDataTableContextMenuEntryInterface } from './data/terra-data-table-context-menu-entry.interface';
+import { TerraBaseData } from '../../../data/terra-base.data';
 
 /**
  * @author mkunze
